Added parsing of the updates meta record JSON
We still need to fix up the logic of the list model but now at least we
can properly consume the JSON returned from the server.

+++ EosAppStore/lib/eos-app-list-model.c
@@ -56,7 +56,7 @@ struct _EosAppListModel
 
   GHashTable *apps;
 
-  guint monotonic_update_id;
+  gint64 monotonic_update_id;
 
   guint applications_changed_id;
   guint changed_guard_id;
